requirements.txt
gradio
langchain
langchain-community
groq
langchain-groq
pypdf
fastembed
faiss-cpu
python-dotenv